Allow any GID to be modified through modify_sid call
modify_sid currently only allows GID=1 to be modified.
However, $HOME_NET, $EXTERNAL_NET and port options are valid use
cases for modifying GID:3. This patch modifies the function so that
any rule can be modified.
Users need to specify the rule ID as gid-sid in modifysid.conf.
In case gid is not specified, it defaults to 1 so as to support
existing configurations. Specifying "*" changes all SIDs for GID-1,
however, specifying 3-* will change all SIDs for GID:3