Solve #313: Join `IPRVersion` path with slash
The `IPRVersion` variable is already checked to ensure it
contains a valid directory. This change just ensures that
the blacklist version is written to
`.../iplists/IPRVersion.dat` rather than
`.../iplistsIPRVersion.dat`.

The added slash should cause no trouble if the `IPRVersion`
variable contains a trailing slash, as repeated slashes are
idempotent.